+++ mobile/lib/db/embeddings_db.dart
